simple IPC event listeners in dedicated file
